Moved library calls to the top, now using derfinder::getTotalMapped()
I loved the library calls to the top so that if a given package is not installed
the job will terminate early. It's a pain to run a bunch of things and then
realize that you were missing some dependencies.

I also switched to derfinder::getTotalMapped() since that function is
publicly available. It's a bit different than the one Andrew Jaffe originally
wrote which is why I adapted the code to run it.